move FiatChangeView
