+++ src/components/display/FiatChangeView.js
