+++ src/screens/Assets/wallet/WalletListItem.js
@@ -25,7 +25,7 @@ import styled from 'styled-components/native';
 // Components
 import Text from 'components/core/Text';
 import FiatValueView from 'components/modern/FiatValueView';
-import FiatChangeView from 'components/modern/FiatChangeView';
+import FiatChangeView from 'components/display/FiatChangeView';
 import TokenIcon from 'components/Icons/TokenIcon';
 
 // Selectors
